# swagger.yaml
openapi: 3.0.1
info:
  title: GPT2 APIs
  description: API Document for serving gpt2 models with torch serve.
  version: 0.3.0
servers:
  - url: https://feature-add-torch-serve-gpt-2-server-gkswjdzz.endpoint.ainize.ai
paths:
  /ping:
    get:
      description: Get TorchServe status.
      operationId: ping
      parameters: []
      responses:
        '200':
          description: TorchServe status
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                properties:
                  status:
                    type: string
                    description: Overall status of the TorchServe.
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - type
                  - message
                properties:
                  code:
                    type: integer
                    description: Error code.
                  type:
                    type: string
                    description: Error type.
                  message:
                    type: string
                    description: Error message.
      deprecated: true
  '/infer/{model_name}':
    post:
      description: Predictions entry point to get inference using default model version.
      parameters:
        - in: path
          name: model_name
          description: Models that can send requests to the server. It may take 15 to 30 seconds on initial request. After that, you can receive requests quickly.
          required: true
          schema:
            type: string
            enum: [
              "GPT2-Colloquial",
              "GPT2-Rick-N-Morty",
              "GPT2-large-Friends",
              "GPT2-large_Fairytale",
              "GPT2-large_LoveCraft",
              "GPT2-large_MyLittlePony",
              "GPT2-large_SuperheroesPower",
              "GPT2-large_SuperheroesStory",
              "GPT2-large_TheOffice",
              "GPT2-large_game_of_thrones",
              "GPT2-large_mcu",
              "GPT2-large_simpsons",
              "GPT2-marketing-man",
              "KoCulture-Large",
              "article_large",
              "cover-letter-gpt2",
              "gpt2-PopSongLyrics",
              "gpt2-Spongebob50",
              "gpt2-bbc",
              "gpt2-large",
              "gpt2-large_Futurama_50",
              "gpt2_barackobama",
              "gpt2_billgates",
              "gpt2_debate",
              "gpt2_dualipa",
              "gpt2_elonmusk",
              "gpt2_film",
              "gpt2_iamcardib",
              "gpt2_joebiden",
              "gpt2_ladygaga",
              "gpt2_stephenking",
              "gpt2_story",
              "gpt2_tim_cook",
              "gpt2_trump"
            ]
            default: "gpt2-large"
      requestBody:
        description: Input data format is defined by each model.
        content:
          'application/json':
            schema:
              type: object
              properties:
                text:
                  type: string
                num_samples:
                  type: integer
                length:
                  type: integer
              example: {
                "text": "Hi There is a Test.",
                "num_samples": 5,
                "length": 20
              }
        required: true
      responses:
        '200':
          description: Output data format is defined by each model.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Model not found or Model Version not found
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - type
                  - message
                properties:
                  code:
                    type: integer
                    description: Error code.
                  type:
                    type: string
                    description: Error type.
                  message:
                    type: string
                    description: Error message.
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - type
                  - message
                properties:
                  code:
                    type: integer
                    description: Error code.
                  type:
                    type: string
                    description: Error type.
                  message:
                    type: string
                    description: Error message.
        '503':
          description: No worker is available to serve request
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - type
                  - message
                properties:
                  code:
                    type: integer
                    description: Error code.
                  type:
                    type: string
                    description: Error type.
                  message:
                    type: string
                    description: Error message.